pyproject.toml
[tool.poetry]
name = "package1"  # Sin espacios
version = "0.1.0"
description = "Proyecto de búsqueda semántica con Flask y PostgreSQL"
authors = ["Javier Cadena <aj-cadena@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
pgvector = "^0.2.5"  # Versión estable
openai = "^1.60.2"
sqlalchemy = ">=2.0,<2.1"
psycopg2-binary = "^2.9.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


